requirements.txt
flask
torch
transformers
pillow
protobuf
sentencepiece
